requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic-settings>=2.0.0
anthropic>=0.39.0
openai>=1.0.0
requests>=2.31.0
xgboost>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Chinese market data sources
tushare>=1.2.89
akshare>=1.10.0
jieba>=0.42.1
